chore: move testutils to test file Signed-off-by: Philippe Martin <phmartin@redhat.com>
